Split up contenttypes_tests.
